move UserProfile component into containers folder
